+++ redGulp/src/main.js
@@ -1,4 +1,17 @@
 var gulp = require('gulp');
+var gutil = require('gulp-util');
+var zip = require("gulp-zip");
+var fs = require("fs");
+var del = require('del');
+var sass = require('gulp-sass');
+var rename = require('gulp-rename');
+var concat = require('gulp-concat');
+var uglify = require('gulp-uglify');
+var path = require("path");
+// XML parser
+var xml2js = require("xml2js");
+var parser = new xml2js.Parser();
+// Get console args
 
 // Load config
 var extension = require('../../package.json');
@@ -135,15 +148,6 @@ gulp.task("release:languages", function () {
 	);
 });
 
-// Overwrite "release" method
-gulp.task("release",
-	gulp.series(
-		"release:plugin",
-		"release:module",
-		"release:redshop"
-	)
-);
-
 gulp.task("release:md5:generate", function () {
 
 	gutil.log(gutil.colors.yellow("Create checksum.md5 file in: checksum.md5"));
@@ -258,10 +262,87 @@ gulp.task("release:md5:clean", gulp.series("release:md5:json"), function () {
 	return gulp.src("./checksum.md5").pipe(clean({ force: true }));
 });
 
+gulp.task("release:redshop",
+	gulp.series("scripts:components.redshop",
+		"sass:components.redshop",
+		"composer:libraries.redshop",
+		"composer:plugins.redshop_pdf.tcpdf")
+	, function (cb) {
+		//fs.readFile("../../../redshop.xml", function (err, data) {
+		//    parser.parseString(data, function (err, result) {
+
+		//    });
+		//});
+
+		var version = result.extension.version[0];
+		var fileName = argv.skipVersion ? "redshop.zip" : "redshop-v" + version + ".zip";
+		var dest = config.releaseDir;
+		console.log('[DEBUG] -------------> ' + dest);
+
+		gutil.log(gutil.colors.grey("===================================================================="));
+		gutil.log(gutil.colors.cyan.bold("redSHOP"), "  |  ", gutil.colors.yellow.bold(version), "  |  ", gutil.colors.white.bold(path.join(config.releaseDir + '/', fileName)));
+		gutil.log(gutil.colors.grey("===================================================================="));
+		var src = getIncludedExtensions();
+		src = src.concat([
+			"./component/**/*",
+			"./component/**/.gitkeep",
+			"./libraries/redshop/**/*",
+			"./libraries/redshop/vendor/**/*",
+			"./libraries/redshop/.gitkeep",
+			'!./**/composer.json',
+			'!./**/composer.lock',
+			'!./**/vendor/**/*.md',
+			'!./**/vendor/**/*.txt',
+			'!./**/vendor/**/*.TXT',
+			'!./**/vendor/**/*.pdf',
+			'!./**/vendor/**/LICENSE',
+			'!./**/vendor/**/CHANGES',
+			'!./**/vendor/**/README',
+			'!./**/vendor/**/VERSION',
+			'!./**/vendor/**/composer.json',
+			'!./**/vendor/**/.gitignore',
+			'!./**/vendor/**/docs',
+			'!./**/vendor/**/docs/**',
+			'!./**/vendor/**/tests',
+			'!./**/vendor/**/tests/**',
+			'!./**/vendor/**/unitTests',
+			'!./**/vendor/**/unitTests/**',
+			'!./**/vendor/**/.git',
+			'!./**/vendor/**/.git/**',
+			'!./**/vendor/**/examples',
+			'!./**/vendor/**/examples/**',
+			'!./**/vendor/**/build.xml',
+			'!./**/vendor/**/phpunit.xml',
+			'!./**/vendor/**/phpunit.xml.dist',
+			'!./**/vendor/**/phpcs.xml',
+			"!./**/vendor/**/Vagrantfile",
+			"./media/**/*",
+			"./media/**/.gitkeep",
+			"!./media/com_redshop/scss",
+			"!./media/com_redshop/scss/**",
+			"./*(install.php|LICENSE.txt|redshop.xml)"
+		]);
+		return gulp.src(src, { base: "./" })
+			.pipe(zip(fileName))
+			.pipe(gulp.dest(dest))
+			.on("end", cb);
+
+		cb();
+	});
+
 gulp.task("release:md5",
 	gulp.series(
 		"release:md5:generate",
 		"release:md5:json",
 		"release:md5:clean"
+	)
+);
+
+// Overwrite "release" method
+gulp.task("release",
+	gulp.series(
+		"release:plugin",
+		"release:module",
+		"release:redshop"
 	)
 );
